Code Smell Fix
The function to render the status menu borders worked, but it was extremely hard to understand and would be virtually impossible for someone else to make changes if needed. The changes makes the code more readable and easier to work with in the future. This fix fell under the "Bloaters" category of code smells.