Code Smell Fix
Fixed a "change preventer" code smell. The hit/miss icon classes were the same thing only with a different image being render. This meant that changes in either class would result in the other class needing changed too. To fix this issue, I made a BoardIcon class with a icon_type parameter. This means that when instantiating the object that either "HIT" or "MISS" needs to be passed in which will render the respective image. This makes understanding and modifying the code in the future easier.